Refactor move the package 'application' outside of the package 'commands'
