+++ commands/archive/archive.go
@@ -10,7 +10,7 @@ import (
 	gp "github.com/simulot/immich-go/adapters/googlePhotos"
 
 	"github.com/simulot/immich-go/adapters/folder"
-	"github.com/simulot/immich-go/commands/application"
+	"github.com/simulot/immich-go/application"
 	"github.com/simulot/immich-go/internal/fileevent"
 	"github.com/simulot/immich-go/internal/filenames"
 	"github.com/simulot/immich-go/internal/fshelper"
