+++ commands/upload/fromGooglePhotos.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	gp "github.com/simulot/immich-go/adapters/googlePhotos"
-	"github.com/simulot/immich-go/commands/application"
+	"github.com/simulot/immich-go/application"
 	"github.com/simulot/immich-go/internal/filenames"
 	"github.com/simulot/immich-go/internal/fshelper"
 	"github.com/spf13/cobra"
