+++ commands/upload/upload.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	"github.com/simulot/immich-go/commands/application"
+	"github.com/simulot/immich-go/application"
 	"github.com/simulot/immich-go/internal/fileevent"
 	"github.com/simulot/immich-go/internal/filters"
 	"github.com/spf13/cobra"
