+++ main.go
@@ -7,8 +7,8 @@ import (
 	"os"
 	"os/signal"
 
+	"github.com/simulot/immich-go/application"
 	"github.com/simulot/immich-go/commands"
-	"github.com/simulot/immich-go/commands/application"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
